UI tweaks and refactoring
UI tweaks and refactoring
