+++ app/src/main/java/com/satti/fs/android/nbv/activities/MainActivity.java
@@ -1,10 +1,11 @@
-package com.satti.fs.android.nbv;
+package com.satti.fs.android.nbv.activities;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+import com.satti.fs.android.nbv.R;
 import com.satti.fs.android.nbv.frag.VenueListingFragment;
 
 public class MainActivity extends AppCompatActivity {
