+++ app/src/main/res/layout/activity_venue_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.satti.fs.android.nbv.VenueDetailActivity">
+    tools:context="com.satti.fs.android.nbv.activities.VenueDetailActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
