+++ app/src/main/res/values/colors.xml
@@ -4,5 +4,5 @@
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
     <color name="background_grey">#f6f6f6</color>
-    <color name="venue_name_color">#4c4c4c</color>
+    <color name="default_text_color">#4c4c4c</color>
 </resources>
