+++ app/src/main/res/values/dimens.xml
@@ -9,6 +9,7 @@
     <dimen name="padding_5">5dp</dimen>
     <dimen name="padding_10">10dp</dimen>
 
+    <dimen name="margin_30">30dp</dimen>
     <dimen name="margin_10">10dp</dimen>
     <dimen name="margin_5">5dp</dimen>
 
@@ -18,6 +19,8 @@
 
 
     <!-- text sizes in sp-->
+    <dimen name="text_size_18">18sp</dimen>
+    <dimen name="text_size_17">17sp</dimen>
     <dimen name="text_size_15">15sp</dimen>
     <dimen name="text_size_12">12sp</dimen>
 </resources>
